api/wayland: move window and monitor to mods.
